[64-BIT] - FULLY WORKING
The 64-bit build is done. It's running, it runs processes, the systimer is ticking away,
there are no crashes because EL1 stack is kept under control.

At this point what can be seen is the 32-bit build is now breaking. It looks like a keyboard
(USPi) related issue. But this now needs to be investigated under its own merits, looking to see
what has changed recently in getting to the 64-bit build working, that has perhaps had an adverse
effect on the 32-bit version.

+++ kernel/c/common/swi/os_setcolour.c
@@ -14,16 +14,17 @@
  * limitations under the License.
  */
 
+#include <stdint.h>
 #include "swi.h"
 #include "video.h"
 
 unsigned int _kernel_swi_os_setcolour( void )
 {
 
-	unsigned int *regs_ptr;
-	unsigned int f, r, g, b;
+	uintptr_t *regs_ptr;
+	uintptr_t f, r, g, b;
 
-	regs_ptr = (unsigned int *) swi_params_ptr_in;
+	regs_ptr = (uintptr_t *) swi_params_ptr_in;
 
 	f = *(regs_ptr+0);
 	r = *(regs_ptr+1);
